+++ rendu/fbecerri/.vscode/settings.json
@@ -1,5 +1,6 @@
 {
     "files.associations": {
-        "array": "cpp"
+        "hash.h": "c",
+        "hotrace.h": "c"
     }
 }
