+++ rendu/fbecerri/include/hotRace.h
@@ -1,11 +1,29 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   hotrace.h                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: fbecerri <fbecerri@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/04/07 09:56:00 by fbecerri          #+#    #+#             */
+/*   Updated: 2021/04/07 10:22:19 by fbecerri         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef HOTRACE_H
 
 # define HOTRACE_H
+
 # include <sys/types.h>
 # include <unistd.h>
 # include <stdio.h>
 # include <stdlib.h>
 # include <stdbool.h>
 # include <string.h>
+
+# include "hash.h"
+# include "ft_string.h"
+# include "io_read.h"
+# include "io_write.h"
 
 #endif
