+++ rendu/fbecerri/src.mk
@@ -1,1 +1,5 @@
 SRCS += src/main.c
+SRCS += src/io_write.c
+SRCS += src/io_read.c
+SRCS += src/string.c
+SRCS += src/hash.c
